requirements.txt
beautifulsoup4==4.13.4
feedparser==6.0.11
python-dotenv==1.1.0
requests==2.32.3
streamlit==1.44.1
streamlit-autorefresh==1.0.1